requirements.txt
faiss-cpu
numpy
sentence-transformers
streamlit
google-generativeai
openai-whisper
yt-dlp
